refactor: separate DAACell component
